% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-inventory.R
\name{clean_invetory_cpe}
\alias{clean_invetory_cpe}
\title{Title}
\usage{
clean_invetory_cpe(df_inventory = data.frame(), df_ner = data.frame())
}
\arguments{
\item{df_inventory}{data.frame}

\item{df_ner}{data.frame}
}
\value{
data.frame
}
\description{
Title
}
