% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nistcpetools.R
\name{cpe2wfn}
\alias{cpe2wfn}
\title{Title}
\usage{
cpe2wfn(df_cpes = parseCPExml(), map = wfn_charmap())
}
\arguments{
\item{df_cpes}{data.frame}

\item{map}{character}
}
\value{
data.frame
}
\description{
Title
}
