% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-inventory.R
\name{cpe_make_title}
\alias{cpe_make_title}
\title{Title}
\usage{
cpe_make_title(df_inventory = getInventory(), verbose = FALSE)
}
\arguments{
\item{df_inventory}{data.frame}

\item{verbose}{logical}
}
\value{
data.frame
}
\description{
Title
}
