% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nistcpetools.R
\name{cpeNERannotate}
\alias{cpeNERannotate}
\title{Annotate CPEs}
\usage{
cpeNERannotate(
  cpes = cpes_etl(),
  vendor = TRUE,
  product = TRUE,
  version = TRUE
)
}
\arguments{
\item{cpes}{data.frame}

\item{vendor}{logical}

\item{product}{logical}

\item{version}{logical}
}
\value{
data.frame
}
\description{
Annotate CPEs
}
