% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nistcpetools.R
\name{cpes_etl}
\alias{cpes_etl}
\title{Title}
\usage{
cpes_etl(xml_path = getLatestdata(), map = wfn_charmap())
}
\arguments{
\item{xml_path}{character}

\item{map}{character}
}
\value{
data.frame
}
\description{
Title
}
