% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nist-ml.R
\name{getCPEchars}
\alias{getCPEchars}
\title{Title}
\usage{
getCPEchars(
  title = TRUE,
  vendor = TRUE,
  product = TRUE,
  version = TRUE,
  cpes = cpe2wfn()
)
}
\arguments{
\item{title}{character}

\item{vendor}{character}

\item{product}{character}

\item{version}{character}

\item{cpes}{data.frame}
}
\value{
character
}
\description{
Title
}
