% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nistcpetools.R
\name{getCPEsample}
\alias{getCPEsample}
\title{Title}
\usage{
getCPEsample(df = cpe2wfn(), num_samples = 5000, randomize = FALSE)
}
\arguments{
\item{df}{data.frame}

\item{num_samples}{numeric}

\item{randomize}{logical}
}
\value{
data.frame
}
\description{
Title
}
