% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nistcpetools.R
\name{getTopProducts}
\alias{getTopProducts}
\title{Title}
\usage{
getTopProducts(df_cpes = cpe2wfn(), num_top = 100)
}
\arguments{
\item{df_cpes}{data.frame}

\item{num_top}{numeric}
}
\value{
data.frame
}
\description{
Title
}
