% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nistcpetools.R
\name{getTopVendors}
\alias{getTopVendors}
\title{Title}
\usage{
getTopVendors(df_cpes_ner = cpe2wfn(), num_top = 100)
}
\arguments{
\item{df_cpes_ner}{data.frame}

\item{num_top}{numeric}
}
\value{
data.frame
}
\description{
Title
}
