% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-inventory.R
\name{inventory_to_ner}
\alias{inventory_to_ner}
\title{Title}
\usage{
inventory_to_ner(df_sccm = data.frame(), verbose = FALSE)
}
\arguments{
\item{df_sccm}{data.frame}

\item{verbose}{logical}
}
\value{
data.frame
}
\description{
Title
}
