% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nist-ml.R
\name{ml_lstm_layers}
\alias{ml_lstm_layers}
\title{Title}
\usage{
ml_lstm_layers(
  input_texts = "",
  target_texts = "",
  input_characters = "",
  target_characters = "",
  num_encoder_tokens = 0,
  num_decoder_tokens = 0,
  max_encoder_seq_length = 0,
  max_decoder_seq_length = 0
)
}
\arguments{
\item{input_texts}{character}

\item{target_texts}{character}

\item{input_characters}{character}

\item{target_characters}{character}

\item{num_encoder_tokens}{numeric}

\item{num_decoder_tokens}{numeric}

\item{max_encoder_seq_length}{numeric}

\item{max_decoder_seq_length}{numeric}
}
\value{
list
}
\description{
Title
}
