% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nist-ml.R
\name{ml_lstm_model_train}
\alias{ml_lstm_model_train}
\title{Title}
\usage{
ml_lstm_model_train(
  encoder_input_data = NA,
  decoder_input_data = NA,
  decoder_target_data = NA,
  encoder_inputs = NA,
  decoder_inputs = NA,
  decoder_outputs = NA,
  optimizer = "rmsprop",
  loss = "categorical_crossentropy",
  batch_size = 32,
  epochs = 100,
  validation_split = 0.2,
  save_path = "models",
  model_name = make.names(paste(gsub("\\\\.", "", OpenRepGrid::randomWords(3)), collapse
    = "_"))
)
}
\arguments{
\item{encoder_input_data}{list}

\item{decoder_input_data}{list}

\item{decoder_target_data}{list}

\item{encoder_inputs}{list}

\item{decoder_inputs}{list}

\item{decoder_outputs}{list}

\item{optimizer}{character}

\item{loss}{character}

\item{batch_size}{numeric}

\item{epochs}{numeric}

\item{validation_split}{numeric}

\item{save_path}{character}

\item{model_name}{character}
}
\value{
character
}
\description{
Title
}
