% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nist-ml.R
\name{ml_lstm_tokenize}
\alias{ml_lstm_tokenize}
\title{Title}
\usage{
ml_lstm_tokenize(
  text,
  input = c("wfn", "cpe", "txt")[sample.int(3, 1)],
  output = c("wfn", "cpe", "txt")[sample.int(3, 1)]
)
}
\arguments{
\item{text}{data.frame , first and second column are used as input and target texts}

\item{input}{logical , default set as TRUE, otherwise text characters will be used}

\item{output}{logical , default set as TRUE, otherwise text characters will be used}
}
\value{
list , with input_texts,target_texts,input_characters,target_characters,num_encoder_tokens,num_decoder_tokens,max_encoder_seq_length,max_decoder_seq_length
}
\description{
Title
}
