% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-inventory.R
\name{ner_to_inventory}
\alias{ner_to_inventory}
\title{Title}
\usage{
ner_to_inventory(df_ner = data.frame(), df_cpes = cpe2wfn())
}
\arguments{
\item{df_ner}{data.frame}

\item{df_cpes}{data.frame}
}
\value{
data.frame
}
\description{
Title
}
