% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nistcpetools.R
\name{str73enc}
\alias{str73enc}
\title{Encode strings to only 73 accepted characters for custom WFN.
\itemize{
\item Replace accents, dieresis, etc to simple ASCII chars
\item Replace tabs with spaces
\item Deal with valid escaped symbols
\item Replace not valid characters with "*"
}}
\usage{
str73enc(name = character(), na_replace = "*")
}
\arguments{
\item{name}{character vector with CPE names}

\item{na_replace}{character, no valid chars will be replaced with "*" by default}
}
\value{
character
}
\description{
Encode strings to only 73 accepted characters for custom WFN.
\itemize{
\item Replace accents, dieresis, etc to simple ASCII chars
\item Replace tabs with spaces
\item Deal with valid escaped symbols
\item Replace not valid characters with "*"
}
}
