% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nist-package.R
\docType{package}
\name{nist-package}
\alias{nist}
\alias{nist-package}
\title{nist: Cybersecurity Standards Tools}
\description{
Toolkit for analyze standards like CPE (Common Platform Enumeration), CVE, CWE and others.
}
\author{
\strong{Maintainer}: Humbert Costas \email{Humbert.Costas@gmail.com} (\href{https://orcid.org/MOTHER-HACK-3R}{ORCID})

}
\keyword{internal}
